requirements.txt
python-telegram-bot==22.0
python-dotenv==1.0.0
pandas==1.5.3
openpyxl==3.1.2
SQLAlchemy==2.0.23
requests>=2.31.0
pyjwt[crypto]>=2.8.0 